handling numpy.int deprecation

+++ pyradigm/tests/test_classify_dataset.py
@@ -151,7 +151,7 @@ def test_set_existing_sample():
 
 def test_data_type():
 
-    for in_dtype in [np.float_, np.int, np.bool_]:
+    for in_dtype in [np.float_, int, np.bool_]:
         cds = ClfDataset(dtype=in_dtype)
         cds.add_samplet('a', [1, 2.0, -434], 'class')
         if cds.dtype != in_dtype or cds['a'].dtype != in_dtype:
